Move files into logical order
